In StreamProcessor, in the nested class MetricProcess: the notifyNewQuad method was inserting elements into the queue by invoking its add method. According to the documentation:
"Inserts the specified element into this queue if it is possible to do so immediately without
violating capacity restrictions, returning true upon success and throwing an
IllegalStateException if no space is currently available".

Hence, it's clear that using this method would cause triples to be lost as soon as the queue
gets filled up to its capacity. Moreover, the problem was successfully reproduced by setting a delay of 1sec
on the computing time of the Blank Metric: it caused an IllegalStateException to be thrown on notifyQuad().

The problem was fixed by invoking the queue's put method instead of add.